style: move fixtures to the file conftest
